move cinder_client_for_test to testtools
