+++ flocker/node/agents/testtools.py
@@ -22,6 +22,7 @@ from twisted.trial.unittest import SynchronousTestCase, SkipTest
 
 from eliot.testing import validate_logging, LoggedAction
 
+from .cinder import authenticated_cinder_client
 from .blockdevice import (
     BlockDeviceDeployer, LoopbackBlockDeviceAPI, IBlockDeviceAPI,
     BlockDeviceVolume, UnknownVolume, AlreadyAttachedVolume,
@@ -576,3 +577,15 @@ def todo_except(supported_tests):
 
         return test_case
     return decorator
+
+
+@require_cinder_credentials
+def cinder_client_for_test(
+        test_case, OPENSTACK_API_USER, OPENSTACK_API_KEY
+):
+    client = authenticated_cinder_client(
+        username=OPENSTACK_API_USER,
+        api_key=OPENSTACK_API_KEY,
+        region='DFW',
+    )
+    return client
